Replace isInt() with isNumber() and get/asInt() with asDouble(), sometimes static_cast back to int/int64_t (#2561)

+++ vnext/ReactUWP/Views/FlyoutViewManager.cpp
@@ -207,7 +207,7 @@ void FlyoutShadowNode::updateProperties(const folly::dynamic&& props)
     {
       if (propertyValue.isNumber())
       {
-        m_targetTag = propertyValue.asInt();
+        m_targetTag = static_cast<int64_t>(propertyValue.asDouble());
         updateTargetElement = true;
       }
       else
